feat: Add Slack utility function for getting thread text
